Inventory Progress
There are still a lot of bugs:
•When you throw the weapon it is still part of the inventory
Solution: A method that removes the weapon from inventory.

•The weapon it's not in the right place
Posible solution: A empty game object inside the player and manually move the weapons so it fits pertfectly with the sprite.

•The secondary weapon is active
Posible solution: a for in start that sets every weapon, except the current one, unactive (Thats to check if the player already has a weapon); When the palyer grabs a new weapon it checks if it is the current weapon, if not it disables itself.