Button and recoin
Before the recoil would rotate the weapon when shooting but, it didn't work as I intended, now I made it so the recoil script rotates the spawn position of the bullets.
It has 3 values, the amount of recoil, the snappiness while shooting and the speed it takes the angle to be 0.

Also a sliding door and a button.
The button is simple:
It has 2 events, for being pressed and released.

The door has 3 public methods that can be called by the button:
Open, close and toggle.