Momentary button and press e to interact
Now instead of the button automatically being press when the player steps on it, they need to press the letter e.
On the trigger enter method the button subscribes to the interact event from the player and unsubscribes from it on the trigger exit.

The momentary button is done. when the player presses the key e tells the observer that it is being pressed and the same when the key is released, but it tell it that it was released.

Now weapons are picked up by pressing a key. It works the same as the button.