Scroll to change weapons
Now the inventory isn't responsible to picking up the weapon or throwing it, instead it is another script called "PickUpWeapon" it's purpose is to make the weapon child of the player and change its position and rotation.
The inventory saves the weapon on a list.

Also now the player can use the mouse scroll to change weapons